Rename datatypes to datetype.
