+++ core/specfem/chunk_element/stress_integrand.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "datatypes/chunk_element_view.hpp"
+#include "datatype/chunk_element_view.hpp"
 #include "enumerations/dimension.hpp"
 #include "enumerations/medium.hpp"
 #include <Kokkos_Core.hpp>
