+++ core/specfem/point/impl/field.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "datatypes/interface.hpp"
+#include "datatype/interface.hpp"
 #include "specfem/data_access.hpp"
 #include <Kokkos_Core.hpp>
 #include <type_traits>
