+++ core/specfem/point/stress.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "datatypes/interface.hpp"
+#include "datatype/interface.hpp"
 #include "enumerations/interface.hpp"
 #include "jacobian_matrix.hpp"
 #include "specfem/data_access.hpp"
