+++ docs/sections/api/index.rst
@@ -29,7 +29,7 @@ Old structure
     enumerations/index
     quadrature/index
     mesh/index
-    datatypes/index
+    datatype/index
     execution/index
     medium/index
     kokkos_kernels/index
