+++ include/datatype/impl/chunk_edge_subview.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "datatypes/point_view.hpp"
+#include "datatype/point_view.hpp"
 #include "enumerations/dimension.hpp"
 #include <Kokkos_Core.hpp>
 
@@ -15,7 +15,7 @@ namespace specfem::datatype::impl {
  *
  * @tparam ViewType The type of the parent view this subview accesses
  */
-template <typename ViewType> struct VectorChunkElementSubview {
+template <typename ViewType> struct VectorChunkEdgeSubview {
   /// Index type from the parent view
   using index_type = typename ViewType::index_type;
   /// Point view type for component access
@@ -33,7 +33,7 @@ template <typename ViewType> struct VectorChunkElementSubview {
    * @param index Reference to the index within the parent view
    */
   KOKKOS_INLINE_FUNCTION
-  VectorChunkElementSubview(ViewType &view, const index_type &index)
+  VectorChunkEdgeSubview(ViewType &view, const index_type &index)
       : view(view), index(index) {}
 
   /**
@@ -48,22 +48,7 @@ template <typename ViewType> struct VectorChunkElementSubview {
       specfem::dimension::type D = index_type::dimension_tag,
       typename std::enable_if_t<D == specfem::dimension::type::dim2, int> = 0>
   KOKKOS_INLINE_FUNCTION constexpr auto &operator()(const int &icomp) {
-    return view(index.ispec, index.iz, index.ix, icomp);
-  }
-
-  /**
-   * @brief Access a specific component of the vector (non-const)
-   *
-   * Specialized for 3D case, providing element access with proper indexing.
-   *
-   * @param icomp Component index to access
-   * @return Reference to the component value
-   */
-  template <
-      specfem::dimension::type D = index_type::dimension_tag,
-      typename std::enable_if_t<D == specfem::dimension::type::dim3, int> = 0>
-  KOKKOS_INLINE_FUNCTION constexpr auto &operator()(const int &icomp) {
-    return view(index.ispec, index.iz, index.iy, index.ix, icomp);
+    return view(index.ispec, index.ipoint, icomp);
   }
 
   /**
@@ -79,23 +64,7 @@ template <typename ViewType> struct VectorChunkElementSubview {
       specfem::dimension::type D = index_type::dimension_tag,
       typename std::enable_if_t<D == specfem::dimension::type::dim2, int> = 0>
   KOKKOS_INLINE_FUNCTION constexpr auto operator()(const int &icomp) const {
-    return view(index.ispec, index.iz, index.ix, icomp);
-  }
-
-  /**
-   * @brief Access a specific component of the vector (const)
-   *
-   * Specialized for 3D case, providing const element access with proper
-   * indexing.
-   *
-   * @param icomp Component index to access
-   * @return Value of the component
-   */
-  template <
-      specfem::dimension::type D = index_type::dimension_tag,
-      typename std::enable_if_t<D == specfem::dimension::type::dim3, int> = 0>
-  KOKKOS_INLINE_FUNCTION constexpr auto operator()(const int &icomp) const {
-    return view(index.ispec, index.iz, index.iy, index.ix, icomp);
+    return view(index.ispec, index.ipoint, icomp);
   }
 
   /**
@@ -124,7 +93,7 @@ template <typename ViewType> struct VectorChunkElementSubview {
  *
  * @tparam ViewType The type of the parent view this subview accesses
  */
-template <typename ViewType> struct TensorChunkElementSubview {
+template <typename ViewType> struct TensorChunkEdgeSubview {
   /// Index type from the parent view
   using index_type = typename ViewType::index_type;
   /// Point view type for tensor component access
@@ -153,14 +122,7 @@ template <typename ViewType> struct TensorChunkElementSubview {
       specfem::dimension::type D = index_type::dimension_tag,
       typename std::enable_if_t<D == specfem::dimension::type::dim2, int> = 0>
   KOKKOS_INLINE_FUNCTION constexpr std::size_t extent(const size_t &i) const {
-    return view.extent(i + 3);
-  }
-
-  template <
-      specfem::dimension::type D = index_type::dimension_tag,
-      typename std::enable_if_t<D == specfem::dimension::type::dim3, int> = 0>
-  KOKKOS_INLINE_FUNCTION constexpr std::size_t extent(const size_t &i) const {
-    return view.extent(i + 4);
+    return view.extent(i + 2);
   }
 
   /**
@@ -174,15 +136,7 @@ template <typename ViewType> struct TensorChunkElementSubview {
       typename std::enable_if_t<D == specfem::dimension::type::dim2, int> = 0>
   KOKKOS_INLINE_FUNCTION constexpr static std::size_t
   static_extent(const size_t &i) {
-    return ViewType::static_extent(i + 3);
-  }
-
-  template <
-      specfem::dimension::type D = index_type::dimension_tag,
-      typename std::enable_if_t<D == specfem::dimension::type::dim3, int> = 0>
-  KOKKOS_INLINE_FUNCTION constexpr static std::size_t
-  static_extent(const size_t &i) {
-    return ViewType::static_extent(i + 4);
+    return ViewType::static_extent(i + 2);
   }
 
   /**
@@ -192,7 +146,7 @@ template <typename ViewType> struct TensorChunkElementSubview {
    * @param index Reference to the index within the parent view
    */
   KOKKOS_INLINE_FUNCTION
-  TensorChunkElementSubview(ViewType &view, const index_type &index)
+  TensorChunkEdgeSubview(ViewType &view, const index_type &index)
       : view(view), index(index) {}
 
   /**
@@ -209,24 +163,7 @@ template <typename ViewType> struct TensorChunkElementSubview {
       typename std::enable_if_t<D == specfem::dimension::type::dim2, int> = 0>
   KOKKOS_INLINE_FUNCTION constexpr auto &operator()(const int &icomp,
                                                     const int &idim) {
-    return view(index.ispec, index.iz, index.ix, icomp, idim);
-  }
-
-  /**
-   * @brief Access a specific tensor component (non-const)
-   *
-   * Specialized for 3D case, providing element access with proper indexing.
-   *
-   * @param icomp Component index
-   * @param idim Dimension index
-   * @return Reference to the tensor component
-   */
-  template <
-      specfem::dimension::type D = index_type::dimension_tag,
-      typename std::enable_if_t<D == specfem::dimension::type::dim3, int> = 0>
-  KOKKOS_INLINE_FUNCTION constexpr auto &operator()(const int &icomp,
-                                                    const int &idim) {
-    return view(index.ispec, index.iz, index.iy, index.ix, icomp, idim);
+    return view(index.ispec, index.ipoint, icomp, idim);
   }
 
   /**
@@ -235,6 +172,7 @@ template <typename ViewType> struct TensorChunkElementSubview {
    * Specialized for 2D case, providing const element access with proper
    * indexing.
    *
+   * @tparam D Dimension tag, defaulted to the index type's dimension tag
    * @param icomp Component index
    * @param idim Dimension index
    * @return Value of the tensor component
@@ -244,25 +182,7 @@ template <typename ViewType> struct TensorChunkElementSubview {
       typename std::enable_if_t<D == specfem::dimension::type::dim2, int> = 0>
   KOKKOS_INLINE_FUNCTION constexpr auto operator()(const int &icomp,
                                                    const int &idim) const {
-    return view(index.ispec, index.iz, index.ix, icomp, idim);
-  }
-
-  /**
-   * @brief Access a specific tensor component (const)
-   *
-   * Specialized for 3D case, providing const element access with proper
-   * indexing.
-   *
-   * @param icomp Component index
-   * @param idim Dimension index
-   * @return Value of the tensor component
-   */
-  template <
-      specfem::dimension::type D = index_type::dimension_tag,
-      typename std::enable_if_t<D == specfem::dimension::type::dim3, int> = 0>
-  KOKKOS_INLINE_FUNCTION constexpr auto operator()(const int &icomp,
-                                                   const int &idim) const {
-    return view(index.ispec, index.iz, index.iy, index.ix, icomp, idim);
+    return view(index.ispec, index.ipoint, icomp, idim);
   }
 
   /**
