+++ include/execution/team_thread_md_range_iterator.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "datatypes/impl/register_array.hpp"
+#include "datatype/impl/register_array.hpp"
 #include "policy.hpp"
 #include "void_iterator.hpp"
 #include <Kokkos_Core.hpp>
