+++ include/kokkos_kernels/impl/compute_source_interaction.tpp
@@ -3,7 +3,7 @@
 #include "specfem/boundary_conditions.hpp"
 #include "specfem/boundary_conditions.tpp"
 #include "specfem/assembly.hpp"
-#include "datatypes/interface.hpp"
+#include "datatype/interface.hpp"
 #include "enumerations/dimension.hpp"
 #include "enumerations/medium.hpp"
 #include "enumerations/wavefield.hpp"
