+++ include/utilities/utilities.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "datatypes/interface.hpp"
+#include "datatype/interface.hpp"
 #include "enumerations/interface.hpp"
 #include "specfem_setup.hpp"
 #include <string>
