+++ tests/unit-tests/point/properties/properties_tests.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "datatypes/simd.hpp"
+#include "datatype/simd.hpp"
 #include "specfem_setup.hpp"
 #include "test_helper.hpp"
 #include <Kokkos_Core.hpp>
